Add files via upload
 Iris-flower-classification
This program applies basic machine learning (classification) concepts on Fisher's Iris Data to predict the species of a new sample of Iris flower.

Introduction
 The iris dataset contains three classes of flowers, Versicolor, Setosa, Virginica, and each class contains 4 features, ‘Sepal length’, ‘Sepal width’, ‘Petal length’, ‘Petal width’. The aim of the iris flower classification is to predict flowers based on their specific features.
The data set consists of 50 samples from each of three species of Iris (Iris setosa, Iris virginica and Iris versicolor).
Four features were measured from each sample (in centimetres):
Length of the sepals
Width of the sepals
Length of the petals
Width of the petals
Working of the iris_decision_tree_classifier

The program takes data from the load_iris() function available in sklearn module.
The program then creates a decision tree based on the dataset for classification.
The user is then asked to enter the four parameters of his sample and prediction about the species of the flower is printed to the user.
Working of the iris_selfmade_KNN

The program takes data from the load_iris() function available in sklearn module.
The program then divides the dataset into training and testing samples in 80:20 ratio randomly using train_test_learn() function available in sklearn module.
The training sample space is used to train the program and predictions are made on the testing sample space.
Accuracy score is then calculated by comparing with the correct results of the training dataset.
So in this project, we learned to train our own supervised machine learning model using Iris Flower Classification Project with Machine Learning. Through this project, we learned about machine learning, data analysis, data visualization, model creation, etc.